requirements.txt
ansible==3.0.0
ansible-base==2.10.6
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==3.1.7
django-prometheus==2.1.0
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
prometheus-client==0.9.0
psycopg2==2.8.6
pycparser==2.20
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
